requirements.txt
aiomysql==0.2.0
annotated-types==0.6.0
anyio==3.7.1
appdata==2.2.1
astroid==3.0.3
bcrypt==4.0.1
beautifulsoup4==4.12.3
blinker==1.6.3
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.2.0
click==8.0.3
colored==1.4.3
cryptography==42.0.2
dill==0.3.8
distlib==0.3.8
ecdsa==0.18.0
fastapi==0.109.2
filelock==3.13.1
Flask==1.0.2
geographiclib==2.0
geopy==2.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==2.8
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.3
mccabe==0.7.0
mysql-connector-python==8.3.0
passlib==1.7.4
pipenv==2023.12.1
platformdirs==4.2.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pylint==3.0.3
PyMySQL==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.7
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.36.3
to-requirements.txt==2.0.7
tomlkit==0.12.3
typing_extensions==4.9.0
urllib3==1.24.3
uvicorn==0.27.0.post1
virtualenv==20.25.0
Werkzeug==3.0.0
